Revert some int_ to intc

+++ phonopy/phonon/mesh.py
@@ -44,7 +44,7 @@ class MeshBase(object):
     ----------
     mesh_numbers: ndarray
         Mesh numbers along a, b, c axes.
-        dtype='int_'
+        dtype='intc'
         shape=(3,)
     qpoints: ndarray
         q-points in reduced coordinates of reciprocal lattice
@@ -52,24 +52,25 @@ class MeshBase(object):
         shape=(ir-grid points, 3)
     weights: ndarray
         Geometric q-point weights. Its sum is the number of grid points.
-        dtype='int_'
+        dtype='intc'
         shape=(ir-grid points,)
     grid_address: ndarray
         Addresses of all grid points represented by integers.
-        dtype='int_'
+        dtype='intc'
         shape=(prod(mesh_numbers), 3)
     ir_grid_points: ndarray
         Indices of irreducibple grid points in grid_address.
-        dtype='int_'
+        dtype='intc'
         shape=(ir-grid points,)
     grid_mapping_table: ndarray
         Index mapping table from all grid points to ir-grid points.
-        dtype='int_'
+        dtype='intc'
         shape=(prod(mesh_numbers),)
     dynamical_matrix: DynamicalMatrix
         Dynamical matrix instance to compute dynamical matrix at q-points.
 
     """
+
     def __init__(self,
                  dynamical_matrix,
                  mesh,
@@ -80,7 +81,7 @@ class MeshBase(object):
                  is_gamma_center=False,
                  rotations=None,  # Point group operations in real space
                  factor=VaspToTHz):
-        self._mesh = np.array(mesh, dtype='int_')
+        self._mesh = np.array(mesh, dtype='intc')
         self._with_eigenvectors = with_eigenvectors
         self._factor = factor
         self._cell = dynamical_matrix.get_primitive()
